+++ src/Element/NetsEasyPaymentElement.php
@@ -2,14 +2,14 @@
 
 namespace Drupal\os2forms_payment\Element;
 
-use Drupal\Core\Render\Element\FormElementBase;
+use Drupal\Core\Render\Element\FormElement;
 
 /**
  * Provides an OS2forms payment element'.
  *
  * @FormElement("os2forms_payment_nets_easy_payment")
  */
-class NetsEasyPaymentElement extends FormElementBase {
+class NetsEasyPaymentElement extends FormElement {
 
   /**
    * {@inheritdoc}