+++ src/Exception/RuntimeException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\os2forms_payment\Exception;
+
+/**
+ * A runtime exception.
+ */
+class RuntimeException extends \Exception {
+
+}